Move non-Qt parts of PyChop to scripts
This partially reverts a previous move of code from scripts to
mantidqtinterfaces. The reason is that Abins2D needs to access parts
of the pychop Python code that do not involve Qt, but this is leading
to test failures which require Qt code and scripts to be
well-separated.

As a possible solution, this commit moves _most_ of pychop into
scripts, but leaves PyChop.py and PyChopGui.py in their current
locations. There might be a tidier way of arranging this, or at least
naming things? For now we are using capitalisation to distinguish
PyChop (Qt-related) from pychop (scripts) which doesn't seem ideal.
