+++ qt/python/mantidqtinterfaces/mantidqtinterfaces/PyChop/PyChopGui.py
@@ -22,7 +22,7 @@ import numpy as np
 import os
 import warnings
 import copy
-from .Instruments import Instrument
+from pychop.Instruments import Instrument
 from qtpy.QtCore import (QEventLoop, Qt, QProcess)  # noqa
 from qtpy.QtWidgets import (QAction, QCheckBox, QComboBox, QDialog, QFileDialog, QGridLayout, QHBoxLayout, QMenu, QLabel,
                             QLineEdit, QMainWindow, QMessageBox, QPushButton, QSizePolicy, QSpacerItem, QTabWidget,
@@ -70,7 +70,7 @@ class PyChopGui(QMainWindow):
         super(PyChopGui, self).__init__(parent)
         if window_flags:
             self.setWindowFlags(window_flags)
-        self.folder = os.path.dirname(sys.modules[self.__module__].__file__)
+        self.folder = os.path.dirname(sys.modules['pychop'].__file__)
         for fname in os.listdir(self.folder):
             if fname.endswith('.yaml'):
                 instobj = Instrument(os.path.join(self.folder, fname))
