+++ scripts/pychop/cncs.yaml
