+++ scripts/pychop/mari.yaml
