+++ scripts/pychop/sequoia.yaml
